Added the base shooting mechanic.
Shooting logic is semi done, animation needs to be added, character change stance when aiming hand flip when aiming (this will be done in animation player), and added the flash effect to the training dummy and changed the animation.

Next thing to do is to finally clean up the attack and shooting animations.