First enemy done
Close combat skeleton done with some currently sketchy code.

When in range goes to attack hitbox appears if player still in it when it finished deals damage.

When getting hit it plays a little animation and takes some health away, (tried to add a flash effect, doesn't work well with the already white skeleton).

When health is 0 plays death animation and after a bit it fades away.