Menus (Option, Paused, Main) implemented, creating hub area, making a merchant, made 2 new weapon sprites,
Finished the 3 menus, in the options we can set the volume, the resolution, fullscreen or not, and the controls can now  be remapped, with the option the reset to the base controls. The general area of the hub is done now I'm making a merchant that will handle the main upgrades, working on the new weapon types, the sprites are done all with their charge animations, currently implementing the 3 main merchant features the weapons and the outfit, weapons are simple, the ui is just a switch to change between them and a the skilltree that still needs to be implemented, also Created a color switching shader with it I can make new outfits without creating more animations just a basic color swaps that changes player stats.