Working skilltree UI
Created a basic skilltree ui, we can set the upgrade cost and the max level of each node, every node reacts to their parent nodes, they lock if the parent is locked or if there aren't any points invested, or if both the parent and the child is maxed out the connection between them will change accordingly.  Also made a mockup of a game state resource that will be store these skills and the players stats like current items and upgrade points.