Created 2 more projectile components
Projectile component 1:
Chain component: Based on the chain stat of the projectile, after hitting an enemy it looks for the next closest enemy and moves toward it, this will happen as many times as the chain stat on the projectile.

Component 2:
A force component it either pulls or pushes enemies in range, within the given range and power.