Almost finished
Added the random upgrade generation and options on level up created the rune stone, which is an envirement item that generated random temporary upgrades when interacted with, reworked the upgrade resources so they can generate stats on the fly.
Only thing that's needed is the boss that is just another elite_type.